[add]] locale.js
